Create view helpers hooking into ASSEMBLE event
Create a new event for the detector ("assemble") to assemble
localised urls. Create two view helpers: one to create a single
localised url for a given route and another view helper to create
a listing of all available languages and the appropriate urls to
the current page in the given language.

Every strategy can support assembling and update their code
accordingly.